Implement simulation of robot odometry.
The robot odometry simulation was implemented as a random walk, where
robot initial robot inputs were simulated as a random uniform
distribution. Subseqent inputs where simulated as a velocity adjustment
about that inital input. This adjust would persist for a random number
of samples, which was also generated as a uniform distribution.