Remove check for invalid subject barcodes.
The measurements corresponding to invalid barcodes should be kept, since
they are valid measurements just with incorrect associations. These
measurements will will have a ground truth value of 0 m and 0 rad, which
will be used by the error calculator to ensure that they do not
contribute to the error statistics.