Add check for invalid range and bearings.
A check was added to ensure that the ranges and bearings associated with
the incorrect barcode association is not included in the error statitics
calculation. This is done because the error statistics are calculated
for error in the sensor measurement, not in the subject association.
Including incorrect associations would cause misleading error statitics.