Change landmark standard deviation to variance values.
The previous landmark values in variance were its standard deviation.
This was done to avoid using a different data structure for housing all
the errro statistics values. However, this was confusing since the
"variance" value would actually refer to its "standard deviation". For
this reason, the landmark standard deviation was squared to make it a
variance value, and then square rooted to change it back for use.