Update methods to use the ID class instead of an integer ID.
The odometry and measurement functions where still using integers for
referencing the agents, so that was changed.

Additionally, the vector containing the number of measurements for all
the robots was updated to a map so that the ID class could be used to extract
the data instead of an integer.